[Chat] Rename HttpFoundation message store to Session message store
