+++ src/chat/src/Bridge/Session/.gitattributes
