+++ src/chat/src/Bridge/Session/.gitignore
