+++ src/chat/src/Bridge/Session/CHANGELOG.md
