+++ src/chat/src/Bridge/Session/LICENSE
